Add json_to_plist ("Property List to JSON") command
This is pretty much the reverse of the json_to_plist command.
It has a few odds like unpredictably behavior for non-native
JSON data types like <date> and <data> and floats
(<float> or <real>) tend to lose precision when being cast
into Python data types (they are not equal), but can be used
greatly for basic tmRessources.

Arguments are directly passed to ``json.dump`` so one can
easily specify ``sortKeys=False`` (or ``"sortKeys": false``)
if the lines should not be sorted and instead written randomly.