Update ``in_one_edit()`` and ``append()`` with optional ``edit`` parameter
This is really handy when you define functions where the edit parameter can be omitted.
You just do ``with in_one_edit(view, edit) as edit:`` and you are safe whatever ``edit``
actually is. Only the view must be valid.
If a valid edit was passed to this function it is not closed!