Parse a so-called "option dict" from the first 3 lines of a file
This allows the specification of a target format from the file itself and makes the
popup asking for the target type redundant in special extension cases (e.g.
".YAML-tmLanguage").
Currently supports ``target_format`` and ``ext`` key. The target format will only be
considered if it has not been passed to the command when calling and the ext
option overrides all other "possible" extensions calculated by the plugin.

Example line:
# [PackageDev] target_format: plist, ext: tmLanguage