Add a not really working rule to Oniguruma RegExp for (?x) extended groups
This was supposed to detect # comments after (?x) has been enabled but it is
not possible to do this throughout the whole "string" (terminated by the first
end-match of the parent match which included the match in question, for example).
Thus, this will only work till the end of the line in which (?x) occurred.