Explicitly dedent scalars when starting with (?x)
This will properly dedent common indentation for (?x) patterns, or
rather most of them. Does not detect if (?x) flag is reverted at some
point, but that's unlikely to happen.

Does not work properly with mixed indentation and can not properly
handle tab-indentation within strings (since the YAML dumper
automatically falls back to double-quoted scalars),
so it is recommended to convert indentation to spaces from within ST
before (in the .tmLanguage file).

Closes #58.