Completely dissolve Lib folder
Gets rid of all `sys.path` modifications.

Sublime Text 2 sets the package's path to the "global source" when it
loads a package's plugin. That means that all modules within the package
are available by their name globally and relative packages *within*
subpackages also work, but ONLY when the file is loaded.

That means we need to move our plist_parser import to the root of the
file, so that it is executed at module load, in order to be able to
locate it at all.