Syntax test completions for end comment marks
End comment marks mostly consist of special characters, so we only
provide completions for when the following tests consists only of these
and whitespace. The main use case for this is when you run
`toggle_comment` on a new line to start a new block-comment, when line
comments aren't available.