refactor syntax test highlighting code
now no longer parses the first line every time it checks whether the current line is a test assertion line - it instead uses a ViewEventListener and remembers the syntax test tokens (and updates them when the file is modified), to make it more efficient. Fixes a bug whereby a file named "syntax_test_[...]" without a well-formed header line would cause a traceback in the console.