Sublime Settings: Color scheme completions are too verbose
I don't find it useful to add the "Package: ..." as description. It is too verbose
and breaks with the "convetion" of displaying the data type on the right.

I actually found adding the package name not very useful as it is not done with themes
and most color schemes I saw right now, have unique name anyway, but ok.