Sublime Settings: Don't ignore empty comment lines
Empty lines within comment blocks are most likely used as visual separator and therefore must not be stripped.

A comment is identified as separator instead of ordinary comment if it starts and ends with '//' and is longer then 2 characters. A single '//' is added as empty line of a comment as well.