Fix crash if view shows unsaved settings file.
If a settings file with correct syntax is not yet saved, the view's
`file_name()` method returns None, which causes `filepath.endswith()`
and `USER_PATH in self.view.file_name()` to raise an AttributeError.

Therefore change __init__() to not create an `self.known_settings`
object for unsaved settings and query for a valid `self.known_settings`
at the very beginning of each event handler method.