Remove backwards matches and use anchors instead
The backwards matches were hard to parse in mind and there's not really
any reason for using them when you can just anchor them to the end.
Performance shouldn't be an issue here, either, and the chances of `$`
matching on the previous line are negligible (and would probably result
in a syntax error).