Use snippets for entries instead of completions
They are the proper tool to use for this kind of automation.
Also remove the `i` and `ii` triggers and instead think of it as "c"
being short for "completions entry".

I might try to unify this to "e" for all resource types later, since
"entry" is shared among all of them.