Replace some scope lookups with filename ones
When we're referring to a specific context within a syntax, we want that
to be the exact one we are thinking about rather than allowing users to
potentially override the selector in their setup and messing things up.

We still use the scope when including an entire syntax, because we don't
care about its exact structure then.