Add menu keys snippets as completions
Adding keys should be possible without typing quotation marks. Therfore the existing "Main Keys" are renamed to "Main Key Names" while a new "Main Keys.sublime-completions" is used to provide a kind of snippet for each known menu key.

It is not as convenient as settings auto-completions, but is a simple fist approach.