Fix argument completions in menu children
The selector was targeted to prevent completions in mapping values, but
this does not work within menus since all children entries are already
a value. Furthermore, they can be arbitrarily nested.

Instead, we just special-case menu files, because all other files are
structured similarly in that mappings are not nested.

Fixes #139.