Fix syntax assignment in syntax_test files
A SYNTAX TEST header can contain a relative path or basename of a syntax file. The later one was not respected before this commit.
This commit modifies the AssignSyntaxTestSyntaxListener in following ways to fix that issue:

1. Plain Text is used as safe default when reading a view's syntax.
2. If the view syntax differs from the SYNTAX TEST header, a sublime-syntax file, matching the basename of the header is searched.
   a) If the list is empty, Plaint Text scope is applied.
   b) If none of the found syntaxes matches the current view's one, the first found syntax is applied.

Note: Even if the HEADER defines a relative path, which does not match the view's syntax file,
      the best matching candidate is applied without respect of the header's path.
      This behavior helps finding the syntax-file if a package was renamed.