Use macro to trigger scope completion on commit
This commit replaces the `PostCompletionsListener` mechanism by a macro.

A key binding for `.` is defined, which is enabled if a completion panel is visible.
It runs a macro file to commit the current completion, add the dot and trigger the next completion.

This way the user can decide, whether to continue with adding the next scope by typing the `.` or stop completion after committing by typing `tab` or `enter` according the the user settings.

This will completely avoid any situation, a dot may be added accidentally.

If we do not want to use a macro file to chain the commands, we could also write a custom command.