Fix Save command not working in snippets of sublime-completions
PackageDev provides a key binding `ctrl+s` to convert snippets.

This key binding overrides the default `save` command.

Therefore a sublime-completions file can't be saved,
if the cursor is within the "contents" key as its value is scoped
as `source.sublime.snippet.embedded` which matches the key binding's
selector.

This commit therefore

1. extends the selector of the keybinding to exclude `source.json`
2. disables the `packagedev_snippet_from_raw_snippet` in `source.json`