Enhance settings tooltip visibility for light color schemes
The CSS of the settings tooltips always blended the --background with white,
which makes in invisible on white bacgkound as Windows does not render a shadow
around tooltips. In order to fix that issue, this commit makes use of ST's
`html.light` and `html.dark` classes to define different colors for light and
dark color schemes.