Fix: more general approach to fix color scheme completions
This commit avoids the global keywords 'rules', 'variables' and 'globals' to be suggested within the 'rules' sequence by adding the ` - meta.mapping meta.sequence` selector to the original solution.

This commit therefore partly reverts the previous commit, which used the special meta scopes to achieve the same result.

This commit is therefore a more general approach which works for new top-level keywords as well.