Add special handling for the first line after push/set
This commit scopes the first line after a push/set statement as
`meta.expect-include-list-or-content`.

The following lines are scoped as `meta.include-list` until a none-empty line with a different indention level is detected or it
does not start with a `-` in case the first line is an include.

If the first line does not contain an include, no special scoping is
performed. The sequence is handled as normal `context_content` then.

Some more test cases were added to ensure correctness of the changes.

Note:

1. By pushing into a dedicated context for includes, the
   `meta.include-list` covers `\n` by design the same way as any
   `meta.block` or `meta.group` does in other syntax definitions.
   The include-list ends with the `\n` of the last include-item.

2. The `meta.flow-sequence.yaml` was removed because:
   a) it does not have any function
   b) it seems dedicated for arrays `[]`
   c) it is not part of the default YAML.sublime-syntax
   d) it does not comply with https://github.com/sublimehq/Packages/issues/421