Improve completions after push/set in sublime-syntax
This commit makes use of the `meta.expect-include-list-or-content`
scope in order to return completions for both, contexts and default
keywords.

The reason for this change is the first line after a `push:` or `set:`
command to probably be a list of included contexts or a normal context.

In order to be able to return a merged list of contexts/keywords, some
changes needed to be made in the `SyntaxDefCompletionsListener`.

1. The `on_query_completions()` method is moved to the top of the class
   as it is a public member. Classes should start with public members
   followed by all protected or private ones.

2. The `on_query_completions()` acts as completion selector only. The
   logic to create the completion lists is moved to dedicated private
   member functions in order to be able to combine them as needed.

3. A new branch was added to `on_query_completions()` to handle the
   `meta.expect-include-list-or-content` scopes and return the
   completions for contexts and keywords all together.