Revert "Fix context completions in multi-line push rules"
This reverts commit 832c2cfa16f042c8bc4fcaaa58ed6857bf3091d8.

This commit probposes to revert 832c2c as it breaks command completions in the first line after a push/set statement.

The first line after a push/set can either be
- the start of a list of included contexts
- or it can contain normal commands like `- match`.

Hence auto-completion must suggest the merged list of available contexts and the normal command completions if the cursor is in the first line.

This does not work in v3.2.2 due to the mentioned commit.

We only see context completions now.