Add support for anonymous syntax-context includes (#218)
The latest Lua.sublime-syntax contains include lists with anonymous
contexts, which are not covered by the recent changes to support
multi-line includes.

Example
=======

- match: foo
  push:
    - - meta_scope: meta
      - match: bar
        scope: keyword
        pop: true
    - context-2
    - context 3
    - - match: baz
        pop: true
    - context-4

This commit adds the required changes to support these kinds of mixed
include lists.

anonymous contexts
==================

If a `- -` is detected within a multi-line-include-list a new
`include_anonymous` context is pushed to the stack. This prevents the
`expect_include_list` context to be popped off too early, which would
end the include-list. The `include_anonymous` keeps on stack as long
as the indention is larger than the indention of the current
include-list.

In order to prevent context references to be included into the
completions as long as the `include_anonymous` is on stack, the
`meta.expect-include...` and `meta.include-...` scopes need to be
cleared.

meta-scope changes
==================

As the `include_anonymous` is included several times, it can clear only
a predefined number of meta scopes. That's why the `meta.expect-include-list-or-content`
can't span the whole first line as indicator for context- and key-completions in conjunction with `meta.include-list.sublime-syntax`.

That's why, it is split into

- `meta.expect-include-list-or-content` from line start up to `-`
- `meta.include-list-or-content` for the rest of the line.

The selector in the `syntax_dev.py` was changed accordingly.

logic changes
=============

Another fix was required in `syntax_dev.py` to support normal keyword
completions within the `include_anonymous` context as the second `-`
needs to be ignored.