Theme/Scheme: Add builtin color completions (#233)
This commit adds all the `--...ish` like colors to the variable completions. The completions are implemented in the color_scheme_dev.py to create consistent results with all the locally defined variables and to work around an issue of the sublime-completions file format, which prevents triggers beginning with none-word characters.