Theme: This commit adds the 'case' property
The `case` property enables themes to modify the rendering of the side bar headings like `Folders` or `Open Files`.

It is implemented as string value by this commit.

As this property is basically an enumeration of "upper", "lower", "title", it should get its own meta scope and a value completion in the future.

The current implementation of the JSON/theme syntax mixture causes all additional meta scopes to be cleared from a string value. A new approach/implementation of string values is needed first to enable completions.