[Snippets] Fix regression with keyword.operator.right
This commit fixes a regression introduced with the last commit because
the `syntax_test_scope_selector.txt` was not realized to be the test
file for `Scope Selector.sublime-syntax`.

Only a space between two selectors is a valid `keyword.operator`.
Therefore it's sufficient to match it only after a `string.unquoted.scope-segment.scope-selector`.

Furthermore we'd not be too restrictive with it. Any number and type of
whitespace between selectors is treated the same way, by ST.