Settings: Fix syntax specific encoding completions
This commit modifies 8b904c60797ae56d598d8516265b701d5dc8648b by
handling encoding completions exactly the same way as `color_scheme`
and `theme` keys.

Therefore the decision to return completions for `default_encoding` and
`fallback_encoding` is made directly in the `_value_completions_for`
method, which is meant to do exactly this.

The if-branches are sorted by the key name rather than logical meaning.
Same for the definition of the called methods.

There is no need to merge comments and default values into the result
if a complete list of values is already available.

The `_known_completions()` method automatically marks the default encoding value.