Syntax: Fix testcase scope completion
This commit fixes an issue which causes wrong selectors to be suggested,
when typing following test cases:

  // <-|
  ## <-|

  Note: | denotes to the caret.

The reason is the caret's column being used to extract scope names
instead of the one from the calculated assertion range. The assertion's
column is the one of the first character of the comment punctuation.

before this commit:

    </style>
    ## <- meta.tag.style.end.html entity.name.tag.style.html

with this commit:

    </style>
    ## <-  meta.tag.style.end.html punctuation.definition.tag.begin.html