Fix syntax test tab dialog appearing on navigating test files using `Goto Anything`.
This commit fixes a situation where navigating syntax test files using
`Goto Anything` can be disruptive. Since the view is loaded and PD throws
the dialog for those files that use tabs in test files, this results in
a bad UX.

To prevent this, we simply skip files that are transiently loaded when
navigating `Goto Anything`.