plugins\file_conversion.py too-many-return-statements
Method run of class had 12 return statements (it is recommended by pylint not to have more than 6).

I extracted methods that contains these statements.
Since a return exit the function and a return in the new function will not exit from the caller, in some cases the result is checked in the caller, which returns if needed.